added new files for project restructuring
adding rebuild files
