+++ json/ROP/ROP_baseCollection.json
@@ -1,0 +1,10 @@
+{
+    "collectionItem0": null, 
+    "collectionItem1": null, 
+    "collectionItem2": null, 
+    "collectionItem3": null, 
+    "collectionItem4": null, 
+    "collectionItem5": null, 
+    "collectionItem6": null, 
+    "collectionItem7": null
+}
