+++ json/menuprefs.json
@@ -1,0 +1,17 @@
+{
+    "OBJ": {
+        "Control": "OBJ_baseCollection.json", 
+        "Default": "OBJ_baseCollection.json", 
+        "Shift": "OBJ_baseCollection.json"
+    }, 
+    "SOP": {
+        "Control": "SOP_baseCollection.json", 
+        "Default": "SOP_baseCollection.json", 
+        "Shift": "SOP_baseCollection.json"
+    }, 
+    "VOP": {
+        "Control": "VOP_baseCollection.json", 
+        "Default": "VOP_baseCollection.json", 
+        "Shift": "VOP_baseCollection.json"
+    }
+}
