--- ne_mm_libs/NE_markingMenu_SOP.json
@@ -1,1 +1,0 @@
-{"menuItem1": {"index": 1, "itemType": "button", "menuObjects": {}, "label": "Attribute Wrangle", "isMenu": false, "command": "cmds.createNode(\"attribwrangle\", False)", "icon": "SOP_attribwrangle"}, "menuItem0": {"index": 0, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem3": {"index": 3, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem2": {"index": 2, "itemType": "menu", "menuObjects": {"submenuItem2": {"index": 2, "itemType": "button", "menuObjects": {}, "label": "Detail Wrangle", "isMenu": false, "command": "cmds.createNode(\"attribwrangle\", False)", "icon": "SOP_attribwrangle"}, "submenuItem0": {"index": 0, "itemType": "button", "menuObjects": {}, "label": "Point Wrangle", "isMenu": false, "command": "cmds.createNode(\"attribwrangle\", False)", "icon": "SOP_attribwrangle"}, "submenuItem1": {"index": 1, "itemType": "button", "menuObjects": {}, "label": "Primitive Wrangle", "isMenu": false, "command": "cmds.createNode(\"attribwrangle\", False)", "icon": "SOP_attribwrangle"}}, "label": "Null", "isMenu": true, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem5": {"index": 5, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem4": {"index": 4, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem7": {"index": 7, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}, "menuItem6": {"index": 6, "itemType": "button", "menuObjects": {}, "label": "Null", "isMenu": false, "command": "cmds.createNode(\"null\", False)", "icon": "SOP_null"}}
