--- ne_mm_libs/__init__.py
@@ -1,1 +1,0 @@
-
