--- ne_mm_libs/collections/OBJ_baseCollection.collection
@@ -1,10 +1,0 @@
-{
-    "collectionItem0": null, 
-    "collectionItem1": null, 
-    "collectionItem2": null, 
-    "collectionItem3": null, 
-    "collectionItem4": null, 
-    "collectionItem5": null, 
-    "collectionItem6": null, 
-    "collectionItem7": null
-}
