--- ne_mm_libs/collections/SOP_baseCollection.collection
@@ -1,98 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"rop_geometry\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "OBJ_geo", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Method Geo Cache", 
-        "menuCollection": "", 
-        "nodetype": "rop_geometry"
-    }, 
-    "collectionItem1": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 1, 
-        "isMenu": true, 
-        "label": "Wrangles", 
-        "menuCollection": "SOP_wrangles.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem2": {
-        "activeWire": false, 
-        "command": "cmds.openEditor()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "COMMON_null", 
-        "index": 2, 
-        "isMenu": false, 
-        "label": "Config", 
-        "menuCollection": "SOP_attributes.collection", 
-        "nodetype": "openEditor()"
-    }, 
-    "collectionItem3": {
-        "activeWire": false, 
-        "command": "cmds.openEditor()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribute", 
-        "index": 3, 
-        "isMenu": true, 
-        "label": "Attributes", 
-        "menuCollection": "SOP_attributes.collection", 
-        "nodetype": "openEditor()"
-    }, 
-    "collectionItem4": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "COMMON_openvdb", 
-        "index": 4, 
-        "isMenu": true, 
-        "label": "Open VDB", 
-        "menuCollection": "SOP_vdbs.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem5": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 5, 
-        "isMenu": true, 
-        "label": "VOPs", 
-        "menuCollection": "SOP_vops.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem6": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_object_merge", 
-        "index": 6, 
-        "isMenu": true, 
-        "label": "Utilities", 
-        "menuCollection": "SOP_soputils.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"null\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "COMMON_null", 
-        "index": 7, 
-        "isMenu": false, 
-        "label": "Null", 
-        "menuCollection": "", 
-        "nodetype": "null"
-    }
-}
