--- ne_mm_libs/collections/SOP_geometry.collection
@@ -1,98 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"grid\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_grid", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Grid", 
-        "menuCollection": "", 
-        "nodetype": "grid"
-    }, 
-    "collectionItem1": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_testgeometry_rubbertoy", 
-        "index": 1, 
-        "isMenu": true, 
-        "label": "Test Geometry", 
-        "menuCollection": "SOP_testgeometry.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem2": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"box\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_box", 
-        "index": 2, 
-        "isMenu": false, 
-        "label": "Box", 
-        "menuCollection": "", 
-        "nodetype": "box"
-    }, 
-    "collectionItem3": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "MISC_python", 
-        "index": 3, 
-        "isMenu": false, 
-        "label": "", 
-        "menuCollection": "", 
-        "nodetype": "null"
-    }, 
-    "collectionItem4": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"sphere\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_sphere", 
-        "index": 4, 
-        "isMenu": false, 
-        "label": "Sphere", 
-        "menuCollection": "", 
-        "nodetype": "sphere"
-    }, 
-    "collectionItem5": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_platonic", 
-        "index": 5, 
-        "isMenu": true, 
-        "label": "Other", 
-        "menuCollection": "SOP_othergeo.collection", 
-        "nodetype": "null"
-    }, 
-    "collectionItem6": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"tube\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_tube", 
-        "index": 6, 
-        "isMenu": false, 
-        "label": "Tube", 
-        "menuCollection": "", 
-        "nodetype": "tube"
-    }, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "", 
-        "commandType": "None", 
-        "context": "SOP", 
-        "icon": "SOP_group", 
-        "index": 7, 
-        "isMenu": true, 
-        "label": "Groups", 
-        "menuCollection": "SOP_groups.collection", 
-        "nodetype": "null"
-    }
-}
