--- ne_mm_libs/collections/SOP_groups.collection
@@ -1,43 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"grouppromote\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_grouppromote", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Group Promote", 
-        "menuCollection": "", 
-        "nodetype": "grouppromote"
-    }, 
-    "collectionItem1": null, 
-    "collectionItem2": null, 
-    "collectionItem3": null, 
-    "collectionItem4": null, 
-    "collectionItem5": null, 
-    "collectionItem6": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"groupdelete\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_groupdelete", 
-        "index": 6, 
-        "isMenu": false, 
-        "label": "Group Delete", 
-        "menuCollection": "", 
-        "nodetype": "groupdelete"
-    }, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"groupcreate\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_group", 
-        "index": 7, 
-        "isMenu": false, 
-        "label": "Group", 
-        "menuCollection": "", 
-        "nodetype": "groupcreate"
-    }
-}
