--- ne_mm_libs/collections/SOP_testgeometry.collection
@@ -1,43 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"testgeometry_pighead\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_testgeometry_pighead", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Pig Head", 
-        "menuCollection": "", 
-        "nodetype": "testgeometry_pighead"
-    }, 
-    "collectionItem1": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"testgeometry_rubbertoy\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_testgeometry_rubbertoy", 
-        "index": 1, 
-        "isMenu": false, 
-        "label": "Rubber Toy", 
-        "menuCollection": "", 
-        "nodetype": "testgeometry_rubbertoy"
-    }, 
-    "collectionItem2": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"testgeometry_squab\", False)", 
-        "commandType": "createnode", 
-        "context": "SOP", 
-        "icon": "SOP_testgeometry_squab", 
-        "index": 2, 
-        "isMenu": false, 
-        "label": "Squab", 
-        "menuCollection": "", 
-        "nodetype": "testgeometry_squab"
-    }, 
-    "collectionItem3": null, 
-    "collectionItem4": null, 
-    "collectionItem5": null, 
-    "collectionItem6": null, 
-    "collectionItem7": null
-}
