--- ne_mm_libs/collections/SOP_vops.collection
@@ -1,65 +1,0 @@
-{
-    "collectionItem0": null, 
-    "collectionItem1": null, 
-    "collectionItem2": null, 
-    "collectionItem3": {
-        "activeWire": false, 
-        "command": "cmds.createVertexVop()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 3, 
-        "isMenu": false, 
-        "label": "Vertex VOP", 
-        "menuCollection": "", 
-        "nodetype": "createVertexVop()"
-    }, 
-    "collectionItem4": {
-        "activeWire": false, 
-        "command": "cmds.createPrimitiveVOP()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 4, 
-        "isMenu": false, 
-        "label": "Primitive VOP", 
-        "menuCollection": "", 
-        "nodetype": "createPrimitiveVOP()"
-    }, 
-    "collectionItem5": {
-        "activeWire": false, 
-        "command": "cmds.createPointVOP()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 5, 
-        "isMenu": false, 
-        "label": "Point VOP", 
-        "menuCollection": "", 
-        "nodetype": "createPointVOP()"
-    }, 
-    "collectionItem6": {
-        "activeWire": false, 
-        "command": "cmds.createVolumeVOP()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 6, 
-        "isMenu": false, 
-        "label": "Volume VOP", 
-        "menuCollection": "", 
-        "nodetype": "createVolumeVOP()"
-    }, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "cmds.createDetailVOP()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribvop", 
-        "index": 7, 
-        "isMenu": false, 
-        "label": "Detail VOP", 
-        "menuCollection": "", 
-        "nodetype": "createDetailVOP()"
-    }
-}
