--- ne_mm_libs/collections/SOP_wrangles.collection
@@ -1,65 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createDetailWrangle()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Detail Wrangle", 
-        "menuCollection": "", 
-        "nodetype": "createDetailWrangle()"
-    }, 
-    "collectionItem1": {
-        "activeWire": false, 
-        "command": "cmds.createPointWrangle()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 1, 
-        "isMenu": false, 
-        "label": "Point Wrangle", 
-        "menuCollection": "", 
-        "nodetype": "createPointWrangle()"
-    }, 
-    "collectionItem2": {
-        "activeWire": false, 
-        "command": "cmds.createPrimitiveWrangle()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 2, 
-        "isMenu": false, 
-        "label": "Primitive Wrangle", 
-        "menuCollection": "", 
-        "nodetype": "createPrimitiveWrangle()"
-    }, 
-    "collectionItem3": null, 
-    "collectionItem4": null, 
-    "collectionItem5": null, 
-    "collectionItem6": {
-        "activeWire": false, 
-        "command": "cmds.createVertexWrangle()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 6, 
-        "isMenu": false, 
-        "label": "Vertex Wrangle", 
-        "menuCollection": "", 
-        "nodetype": "createVertexWrangle()"
-    }, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "cmds.createVolumeWrangle()", 
-        "commandType": "customFunction", 
-        "context": "SOP", 
-        "icon": "SOP_attribwrangle", 
-        "index": 7, 
-        "isMenu": false, 
-        "label": "Volume Wrangle", 
-        "menuCollection": "", 
-        "nodetype": "createVolumeWrangle()"
-    }
-}
