--- ne_mm_libs/collections/VOP_basicmath.collection
@@ -1,76 +1,0 @@
-{
-    "collectionItem0": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"complement\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_complement", 
-        "index": 0, 
-        "isMenu": false, 
-        "label": "Complement", 
-        "menuCollection": "", 
-        "nodetype": "complement"
-    }, 
-    "collectionItem1": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"add\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_add", 
-        "index": 1, 
-        "isMenu": false, 
-        "label": "Add", 
-        "menuCollection": "", 
-        "nodetype": "add"
-    }, 
-    "collectionItem2": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"multiply\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_multiply", 
-        "index": 2, 
-        "isMenu": false, 
-        "label": "Multiply", 
-        "menuCollection": "", 
-        "nodetype": "multiply"
-    }, 
-    "collectionItem3": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"divide\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_divide", 
-        "index": 3, 
-        "isMenu": false, 
-        "label": "Divide", 
-        "menuCollection": "", 
-        "nodetype": "divide"
-    }, 
-    "collectionItem4": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"constant\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_constant", 
-        "index": 4, 
-        "isMenu": false, 
-        "label": "Constant", 
-        "menuCollection": "", 
-        "nodetype": "constant"
-    }, 
-    "collectionItem5": null, 
-    "collectionItem6": null, 
-    "collectionItem7": {
-        "activeWire": false, 
-        "command": "cmds.createNode(\"maddconst\", False)", 
-        "commandType": "createnode", 
-        "context": "VOP", 
-        "icon": "VOP_maddconst", 
-        "index": 7, 
-        "isMenu": false, 
-        "label": "Multiply-Add Constant", 
-        "menuCollection": "", 
-        "nodetype": "maddconst"
-    }
-}
